requirements.txt
strands-agents
strands-agents-tools
mcp
pygithub
python-dotenv
requests
ollama
anthropic
boto3
streamlit